+++ Example/Emission/ARRootViewController.m
@@ -277,7 +277,7 @@
 - (ARCellData *)jumpToArtworkInAuction
 {
   return [self tappableCellDataWithTitle:@"Artwork In Auction" selection:^{
-    id viewController = [[ARArtworkComponentViewController alloc] initWithArtworkID:@"marina-donati-insects"];
+    id viewController = [[ARArtworkComponentViewController alloc] initWithArtworkID:@"italy-tuscany-14th-century-self-portrait"];
     [self.navigationController pushViewController:viewController animated:YES];
   }];
 }