+++ src/__generated__/CommercialInformation_artwork.graphql.ts
@@ -1,6 +1,8 @@
 /* tslint:disable */
 
 import { ReaderFragment } from "relay-runtime";
+import { ArtworkExtraLinks_artwork$ref } from "./ArtworkExtraLinks_artwork.graphql";
+import { AuctionCountDownTimer_artwork$ref } from "./AuctionCountDownTimer_artwork.graphql";
 import { CommercialButtons_artwork$ref } from "./CommercialButtons_artwork.graphql";
 import { CommercialEditionSetInformation_artwork$ref } from "./CommercialEditionSetInformation_artwork.graphql";
 import { CommercialPartnerInformation_artwork$ref } from "./CommercialPartnerInformation_artwork.graphql";
@@ -10,20 +12,23 @@ export type CommercialInformation_artwork = {
     readonly availability: string | null;
     readonly artists: ReadonlyArray<{
         readonly isConsignable: boolean | null;
-        readonly name: string | null;
     } | null> | null;
     readonly editionSets: ReadonlyArray<{
         readonly isAcquireable: boolean | null;
         readonly isOfferable: boolean | null;
         readonly saleMessage: string | null;
     } | null> | null;
+    readonly sale: {
+        readonly isClosed: boolean | null;
+    } | null;
     readonly saleMessage: string | null;
     readonly shippingInfo: string | null;
     readonly shippingOrigin: string | null;
     readonly isAcquireable: boolean | null;
     readonly isOfferable: boolean | null;
     readonly isInquireable: boolean | null;
-    readonly " $fragmentRefs": CommercialButtons_artwork$ref & CommercialPartnerInformation_artwork$ref & CommercialEditionSetInformation_artwork$ref;
+    readonly isInAuction: boolean | null;
+    readonly " $fragmentRefs": CommercialButtons_artwork$ref & CommercialPartnerInformation_artwork$ref & CommercialEditionSetInformation_artwork$ref & AuctionCountDownTimer_artwork$ref & ArtworkExtraLinks_artwork$ref;
     readonly " $refType": CommercialInformation_artwork$ref;
 };
 
@@ -80,13 +85,6 @@ return {
           "name": "isConsignable",
           "args": null,
           "storageKey": null
-        },
-        {
-          "kind": "ScalarField",
-          "alias": null,
-          "name": "name",
-          "args": null,
-          "storageKey": null
         }
       ]
     },
@@ -102,6 +100,24 @@ return {
         (v0/*: any*/),
         (v1/*: any*/),
         (v2/*: any*/)
+      ]
+    },
+    {
+      "kind": "LinkedField",
+      "alias": null,
+      "name": "sale",
+      "storageKey": null,
+      "args": null,
+      "concreteType": "Sale",
+      "plural": false,
+      "selections": [
+        {
+          "kind": "ScalarField",
+          "alias": null,
+          "name": "isClosed",
+          "args": null,
+          "storageKey": null
+        }
       ]
     },
     (v2/*: any*/),
@@ -129,6 +145,13 @@ return {
       "storageKey": null
     },
     {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "isInAuction",
+      "args": null,
+      "storageKey": null
+    },
+    {
       "kind": "FragmentSpread",
       "name": "CommercialButtons_artwork",
       "args": null
@@ -142,9 +165,19 @@ return {
       "kind": "FragmentSpread",
       "name": "CommercialEditionSetInformation_artwork",
       "args": null
+    },
+    {
+      "kind": "FragmentSpread",
+      "name": "AuctionCountDownTimer_artwork",
+      "args": null
+    },
+    {
+      "kind": "FragmentSpread",
+      "name": "ArtworkExtraLinks_artwork",
+      "args": null
     }
   ]
 };
 })();
-(node as any).hash = '83b5769837fc412de226dfda25c47ca7';
+(node as any).hash = 'f71aa77db016d3a64418773480a63eab';
 export default node;